Fix erroneous error when leaving VC
When a user uses the 'leave' command, the bot would say there was
an error leaving the voice channel. The bot would actually leave,
it just wouldn't parse the number of deleted rows from redis
properly. Instead, just don't even wait for a reply from redis, as
it's not needed.